rename: 불명확한 폴더 이름 수정
