+++ src/components/pages/main/group-study-list-page/group-hashtag-box/GroupHashtagBox.style.js
