+++ src/components/pages/main/group-study-list-page/group-study-list-page-layout/GroupStudyListPageLayout.style.js
