+++ src/components/pages/main/group-study-list-page/group-study-thumbnail-swiper/GroupStudyLikeThumbnailSwiper.style.js
