+++ src/components/pages/main/group-study-list-page/group-study-thumbnail/GroupStudyLikeThumbnail.style.js
