+++ src/components/pages/main/group-study-list-page/group-study-thumbnail/GroupStudyNewThumbnail.style.js
